+++ sources/plug_osg/include/luna/wrappers/wrapper_osgViewer_OpenGLQuerySupport.h
@@ -15,8 +15,8 @@ public:
 
 	wrapper_osgViewer_OpenGLQuerySupport(lua_State* L, lua_Table* dum) : osgViewer::OpenGLQuerySupport(), luna_wrapper_base(L) {};
 
-	// void osgViewer::OpenGLQuerySupport::checkQuery(osg::Stats * stats, osg::State * state, unsigned long long startTick)
-	void checkQuery(osg::Stats * stats, osg::State * state, unsigned long long startTick) {
+	// void osgViewer::OpenGLQuerySupport::checkQuery(osg::Stats * stats, osg::State * state, __int64 startTick)
+	void checkQuery(osg::Stats * stats, osg::State * state, __int64 startTick) {
 		THROW_IF(!_obj.pushFunction("checkQuery"),"No implementation for abstract function osgViewer::OpenGLQuerySupport::checkQuery");
 		_obj.pushArg(stats);
 		_obj.pushArg(state);
@@ -39,8 +39,8 @@ public:
 		return (_obj.callFunction<void>());
 	};
 
-	// void osgViewer::OpenGLQuerySupport::initialize(osg::State * state, unsigned long long startTick)
-	void initialize(osg::State * state, unsigned long long startTick) {
+	// void osgViewer::OpenGLQuerySupport::initialize(osg::State * state, __int64 startTick)
+	void initialize(osg::State * state, __int64 startTick) {
 		if(_obj.pushFunction("initialize")) {
 			_obj.pushArg(state);
 			_obj.pushArg(startTick);
